move slice() to plexscoring
